+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## Bugfix
 * [#27](https://github.com/epaew/eslint-plugin-filenames-simple/pull/27) Fixed the behavior that ESLint execution fails
 when the lint target includes TypeScript notation and the rule `named-export` is enabled.
+* [#30](https://github.com/epaew/eslint-plugin-filenames-simple/pull/30)
+    * Fixed false positives when statement contains both default export and single named export.
+    * Fixed false positives when statement contains both all export (`export *`) and single named export.
 
 # 0.3.0
 ## Features